+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.20.0 (unreleased)
+
+FEATURES
+
+### Studio
+
+- Add `Run Subflow` widget
+- Add `IncomingParent` transition to the trigger widget
+
 ## v0.19.0 (2021-09-05)
 
 FEATURES